feat: Change comment reference icons
