+++ client/src/components/atoms/CommentReference.vue
@@ -1,0 +1,63 @@
+<template>
+  <q-btn
+    :aria-label="$t(`submissions.comment.reference.go_to_highlight`)"
+    dense
+    flat
+    color="primary"
+    class="q-mr-xs"
+    no-caps
+    @click="setActive"
+  >
+    <q-icon
+      v-if="comment.read_at === null"
+      size="xs"
+      :name="read_name[props.comment.__typename]"
+      color="primary"
+    ></q-icon>
+    <q-icon
+      v-else
+      size="xs"
+      :name="unread_name[props.comment.__typename]"
+      color="warning"
+    ></q-icon>
+    <q-tooltip>{{
+      $t(`submissions.comment.reference.go_to_highlight`)
+    }}</q-tooltip>
+  </q-btn>
+</template>
+
+<script setup>
+import { inject, nextTick } from "vue"
+
+const props = defineProps({
+  comment: {
+    type: Object,
+    required: true,
+  },
+})
+
+const unread_name = {
+  OverallComment: "mark_unread_chat_alt",
+  OverallCommentReply: "mark_unread_chat_alt",
+  InlineComment: "mark_chat_unread",
+  InlineCommentReply: "mark_chat_unread",
+}
+
+const read_name = {
+  OverallComment: "chat",
+  OverallCommentReply: "chat",
+  InlineComment: "chat_bubble",
+  InlineCommentReply: "chat_bubble",
+}
+
+const activeComment = inject("activeComment")
+
+function setActive() {
+  //Null the active comment first to trigger the scroll watcher
+  //TODO: Do this in a more elegant way.
+  activeComment.value = null
+  nextTick(() => {
+    activeComment.value = props.comment
+  })
+}
+</script>
